ci: extract wait_until_done and fix ci
