+++ sdk/ml/azure-ai-ml/tests/command_job/e2etests/test_command_job.py
@@ -1,10 +1,9 @@
-import time
 from pathlib import Path
 from time import sleep
 from typing import Callable
 
 from azure.ai.ml.entities import AmlTokenConfiguration
-from devtools_testutils import AzureRecordedTestCase, is_live, set_bodiless_matcher
+from devtools_testutils import AzureRecordedTestCase, is_live
 import jwt
 import pytest
 
@@ -18,9 +17,10 @@ from azure.ai.ml.entities._job.command_job import CommandJob
 from azure.ai.ml.entities._job.distribution import MpiDistribution
 from azure.ai.ml.entities._job.job import Job
 from azure.ai.ml.exceptions import ValidationException
-from azure.ai.ml.operations._job_ops_helper import _wait_before_polling
 from azure.ai.ml.operations._run_history_constants import JobStatus, RunHistoryConstants
 from azure.core.polling import LROPoller
+
+from test_utilities.utils import wait_until_done
 
 # These params are logged in ..\test_configs\python\simple_train.py. test_command_job_with_params asserts these parameters are
 # logged in the training script, so any changes to parameter logging in simple_train.py must preserve this logging or change it both
@@ -342,19 +342,13 @@ class TestCommandJob(AzureRecordedTestCase):
     def test_command_job_download(self, tmp_path: Path, randstr: Callable[[], str], client: MLClient) -> None:
         client: MLClient = client
 
-        def wait_until_done(job: Job) -> None:
-            poll_start_time = time.time()
-            while job.status not in RunHistoryConstants.TERMINAL_STATUSES:
-                time.sleep(_wait_before_polling(time.time() - poll_start_time))
-                job = client.jobs.get(job.name)
-
         job = client.jobs.create_or_update(
             load_job(
                 source="./tests/test_configs/command_job/command_job_quick_with_output.yml",
                 params_override=[{"name": randstr("name")}],
             )
         )
-        wait_until_done(job)
+        wait_until_done(client=client, job=job)
 
         client.jobs.download(name=job.name, download_path=tmp_path, all=True)
 
@@ -371,12 +365,6 @@ class TestCommandJob(AzureRecordedTestCase):
     def test_command_job_local_run_download(self, tmp_path: Path, randstr: Callable[[], str], client: MLClient) -> None:
         client: MLClient = client
 
-        def wait_until_done(job: Job) -> None:
-            poll_start_time = time.time()
-            while job.status not in RunHistoryConstants.TERMINAL_STATUSES:
-                time.sleep(_wait_before_polling(time.time() - poll_start_time))
-                job = client.jobs.get(job.name)
-
         job = client.jobs.create_or_update(
             load_job(
                 source="./tests/test_configs/command_job/command_job_quick_with_output.yml",
@@ -384,7 +372,7 @@ class TestCommandJob(AzureRecordedTestCase):
             )
         )
 
-        wait_until_done(job)
+        wait_until_done(client=client, job=job)
 
         client.jobs.download(name=job.name, download_path=tmp_path, all=True)
